Port game source files to SDL3 (everything except LAN networking)
Mechanical renames applied tree-wide: SDL_FillRect -> SDL_FillSurfaceRect,
SDL_FreeSurface -> SDL_DestroySurface, event type constants (SDL_QUIT ->
SDL_EVENT_QUIT etc.), SDLKey/SDLMod -> SDL_Keycode/SDL_Keymod, keysym.sym/
mod -> key.key/mod, single-letter SDLK_x -> SDLK_X, SDLK_KP0/9 ->
SDLK_KP_0/9, SDL_ShowCursor(0/1) -> SDL_HideCursor()/SDL_ShowCursor(),
SDL_WM_GrabInput -> SDL_SetWindowMouseGrab, SDL_UpdateRect(s)/SDL_Flip ->
T4K_UpdateRect (present is always full-frame now), SDL_MapRGB argument
list, mouse wheel handled as its own event type instead of a button.

Per-file fixes:
- highscore.c: name-entry text field now uses SDL_EVENT_TEXT_INPUT
  (UTF-8) instead of the removed key event unicode field; needs
  SDL_StartTextInput()/SDL_StopTextInput().
- setup.c/factoroids.c: SDL_MouseButtonEvent.state -> .down;
  SDL_WarpMouse -> SDL_WarpMouseInWindow; SDL_EventState ->
  SDL_SetEventEnabled; SDL_JOY* event constants renamed.
- titlescreen.c/campaign.c: SDL_Surface no longer has a clip_rect
  member (it always equaled the full surface here, since nothing in
  this codebase narrows it) - replaced with ->w/->h or NULL rects.
- comets.c: two lines were calling the gettext macro on wide-char
  string literals, which never actually worked correctly even under
  real gettext (it's a narrow-char API) - dropped the bogus _() calls.
- T4K_Throttle's timer argument is Uint64* now (SDL_GetTicks() returns
  Uint64), so local  variables across several files changed
  from Uint32.
- Fixed a handful of missing #includes and prototype/definition
  signature mismatches (comets.h, get_scale(), HandleTitleScreenEvents/
  ResSwitch) that only became hard errors under a modern compiler.

Verified: everything now compiles cleanly via CMake except menu_lan.c/
comets.c's LAN_* calls and network.c/server.c themselves, which need a
real port to SDL3_net's completely redesigned NET_* API (NET_Address/
NET_StreamSocket/NET_Server/NET_DatagramSocket replace SDLNet_TCPsocket/
UDPsocket/UDPpacket/socket sets) - not yet started.

+++ src/network.c
@@ -78,7 +78,7 @@ int LAN_DetectServers(void)
     int attempts = 0;
     int num_servers = 0;
     int i = 0;
-    Uint32 timer = 0;
+    Uint64 timer = 0;
     //zero out old server list
     for(i = 0; i < MAX_SERVERS; i++)
         servers[i].ip.host = 0;
